removed gulp build because of useless complexity
