+++ lib/Constants.js
@@ -1,1 +1,28 @@
-const path=require("path"),package_json=require(path.join("..","package.json"));module.exports={PRINT_BROWSER_CONSOLE:!1,DROP_KEYFRAMES:!0,PUPPETEER_HEADLESS:!0,BROWSER_USER_AGENT:"Crittr "+package_json.version,BROWSER_CACHE_ENABLED:!0,BROWSER_JS_ENABLED:!0,BROWSER_CONCURRENT_TABS:10,DEVICE_WIDTH:1200,DEVICE_HEIGHT:1080,DEVICE_SCALE_FACTOR:1,DEVICE_IS_MOBILE:!1,DEVICE_HAS_TOUCH:!1,DEVICE_IS_LANDSCAPE:!1,TIMEOUT:3e4,PAGE_LOAD_TIMEOUT:2e3,PAGE_RENDER_TIMEOUT:300,PAGE_SCREENSHOT:!1,OUTPUT_REMAINING_CSS:!0,RULE_SEPARATOR:"-#-"};
+const path         = require('path');
+const package_json = require(path.join("..", "package.json"));
+
+module.exports = {
+    // CRITTR BASED
+
+    PRINT_BROWSER_CONSOLE:   false,
+    DROP_KEYFRAMES:          true,
+    PUPPETEER_HEADLESS:      true,
+    BROWSER_USER_AGENT:              "Crittr " + package_json.version,
+    BROWSER_CACHE_ENABLED:   true,
+    BROWSER_JS_ENABLED:      true,
+    BROWSER_CONCURRENT_TABS: 10,
+    DEVICE_WIDTH:            1200,
+    DEVICE_HEIGHT:           1080,
+    DEVICE_SCALE_FACTOR:     1,
+    DEVICE_IS_MOBILE:        false,
+    DEVICE_HAS_TOUCH:        false,
+    DEVICE_IS_LANDSCAPE:     false,
+    TIMEOUT:                 30000,
+    PAGE_LOAD_TIMEOUT:       2000,
+    PAGE_RENDER_TIMEOUT:     300,
+    PAGE_SCREENSHOT:         false,
+    OUTPUT_REMAINING_CSS:    true,
+
+    // CODE BASED
+    RULE_SEPARATOR: "-#-"
+};
