+++ lib/evaluation/extract_critical_with_css.js
@@ -1,1 +1,223 @@
-module.exports=({sourceAst:e,loadTimeout:t,keepSelectors:s,removeSelectors:r})=>new Promise((n,o)=>{const l=["media","rule"],i=new RegExp(["after","before","first-line","first-letter","selection","visited"].map(e=>":?:"+e).reduce((e,t)=>e+"|"+t),"g"),c=new RegExp(["root"].map(e=>":?:"+e).reduce((e,t)=>e+"|"+t),"g"),u=new RegExp(/:?:-[a-z-]*/g);s=s||[],r=r||[],t=t||2e3;const a=window.innerHeight,d=new Set,p=new Map,f=(e,t)=>{window.requestAnimationFrame(()=>{Date.now()-e>=t?window.stop():f(e,t)})};f(Date.now(),t);const m=e=>{if(y(e))return!0;if(E(e))return!1;const t=g(e);let s;try{s=document.querySelectorAll(t)}catch(e){return!1}const r=s.length;for(let e=0;e<r;e++)if(x(s[e]))return!0;return!1},g=e=>(e.indexOf(!1)&&(e=e.replace(i,"")),e.indexOf(!1)&&(e=e.replace(u,"")),e.indexOf(!1)&&(e=e.replace(c,"")),e),w=e=>e.startsWith(":")&&null===e.match(c),h=e=>(e="^"+e.replace(/([.*><+~])/g,"\\$1").replace(/%/g,".*")+"$",new RegExp(e,"gm")),y=e=>s.includes(e)||s.some(t=>h(t).test(e)),E=e=>r.includes(e)||r.some(t=>h(t).test(e)),x=e=>{if(d.has(e))return!0;return!!(e.getBoundingClientRect().top<a)&&(d.add(e),!0)},R=e=>{let t={},s=null;e.stylesheet?t=e.stylesheet:e.rules&&"media"===e.type?(t=e,s=e.media):console.warn("Missing ast rules!!!",e.type,e.stylesheet);const r=t.rules;for(let e of r)if(l.includes(e.type))if("media"===e.type)R(e);else{const t=e.selectors||[];s=s||"";const r=s+t.join();for(let n of t)if(w(n)||m(n))if(p.has(r)){const e=p.get(r);e.selectors.includes(n)||e.selectors.push(n)}else p.set(r,{selectors:[n],media:s,rule:e})}else console.debug("DEBUG: UNPROCESSED RULE TYPE: "+e.type)};return R(e),n([...p])}).catch(e=>{console.error(e)});
+/**
+ * Used to extract critical css with the help of a source css. This will result in larger size because every vendor
+ * prefix is used.
+ *
+ * @param sourceAst
+ * @param renderTimeout
+ * @param keepSelectors
+ * @returns {Promise<Map<Object>>}
+ */
+module.exports = ({sourceAst, loadTimeout, keepSelectors, removeSelectors}) => {
+    return new Promise((resolve, reject) => {
+        // PRE CONFIG VARS
+        const usedSelectorTypes = [
+            "media",
+            "rule"
+        ];
+
+        const pseudoSelectors = [
+            "after",
+            "before",
+            "first-line",
+            "first-letter",
+            "selection",
+            "visited"
+        ];
+
+        const pseudoExcludes = [
+            "root"
+        ];
+
+        const PSEUDO_DEFAULT_REGEX  = new RegExp(pseudoSelectors.map(s => ":?:" + s).reduce((acc, cur) => acc + "|" + cur), "g");
+        const PSEUDO_EXCLUDED_REGEX = new RegExp(pseudoExcludes.map(s => ":?:" + s).reduce((acc, cur) => acc + "|" + cur), "g");
+        const PSEUDO_BROWSER_REGEX  = new RegExp(/:?:-[a-z-]*/g);
+
+        // ADJUSTMENTS
+        keepSelectors   = keepSelectors || [];
+        removeSelectors = removeSelectors || [];
+        loadTimeout     = loadTimeout || 2000;
+
+        // innerHeight of window to determine if in viewport
+        const height = window.innerHeight;
+
+        // Nodes in above the fold content
+        const criticalNodes     = new Set();
+        // Final result Map
+        const criticalSelectors = new Map();
+
+        const stopPageLoadAfterTimeout = (start, timeout) => {
+            window.requestAnimationFrame(() => {
+                const timePassed = Date.now() - start;
+                if (timePassed >= timeout) {
+                    window.stop();
+                } else {
+                    stopPageLoadAfterTimeout(start, timeout);
+                }
+            })
+        };
+        stopPageLoadAfterTimeout(Date.now(), loadTimeout);
+
+        const isSelectorCritical = (selector) => {
+            if (isSelectorForceIncluded(selector)) return true;
+            if (isSelectorForceExcluded(selector)) return false;
+
+            // clean selector from important pseudo classes to get him tracked as critical
+            const cleanedSelector = getCleanedSelector(selector);
+
+            let elements;
+            try {
+                elements = document.querySelectorAll(cleanedSelector)
+            } catch (e) {
+                // Selector not valid
+                return false
+            }
+
+            // selector has > 0 elements matching -> check for above the fold - break on success
+            const elemLength = elements.length;
+            for (let i = 0; i < elemLength; i++) {
+                if (isElementAboveTheFold(elements[i])) {
+                    return true
+                }
+            }
+            return false
+        };
+
+        /**
+         * Clean selector of pseudo classes
+         *
+         * @param selector
+         * @returns selector {String}
+         */
+        const getCleanedSelector = selector => {
+            // We wont clean selectors without ":" because its faster as to replace all
+            if (selector.indexOf(":" > -1)) {
+                selector = selector.replace(PSEUDO_DEFAULT_REGEX, "");
+            }
+            // Remove browser pseudo selectors
+            if (selector.indexOf(":" > -1)) {
+                selector = selector.replace(PSEUDO_BROWSER_REGEX, "");
+            }
+            // Remove excluded pseudo selectors
+            if (selector.indexOf(":" > -1)) {
+                selector = selector.replace(PSEUDO_EXCLUDED_REGEX, "");
+            }
+
+            return selector;
+        };
+
+        /**
+         * If selector is purely pseudo (f.e. ::-moz-placeholder) -> KEEP IT.
+         * But don't keep excludedPseudos by default
+         *
+         * @param selector
+         * @returns {boolean}
+         */
+        const isPurePseudo = selector => selector.startsWith(":") && selector.match(PSEUDO_EXCLUDED_REGEX) === null;
+
+        /**
+         * Creates a regex out of a wildcard selector. Returns the normal regex for a non wildcard selector
+         *
+         * @param {string} selector
+         * @returns {RegExp} {RegExp}
+         */
+        const getRegexOfSelector = selector => {
+            selector = "^" + selector.replace(/([.*><+~])/g, "\\$1").replace(/%/g, ".*") + "$";
+            return new RegExp(selector, "gm");
+        };
+
+        const isSelectorForceIncluded = selector => {
+            return keepSelectors.includes(selector) || keepSelectors.some( tmpSel => {
+                const selectorWcRegex = getRegexOfSelector(tmpSel); // transform wildcards into regex
+                return selectorWcRegex.test(selector);
+            });
+        };
+
+        const isSelectorForceExcluded = selector => {
+            return removeSelectors.includes(selector) || removeSelectors.some( tmpSel => {
+                const selectorWcRegex = getRegexOfSelector(tmpSel); // transform wildcards into regex
+                return selectorWcRegex.test(selector);
+            });
+        };
+
+        const isElementAboveTheFold = (element) => {
+            if (criticalNodes.has(element)) return true;
+
+            const isAboveTheFold = element.getBoundingClientRect().top < height;
+
+            if (isAboveTheFold) {
+                criticalNodes.add(element);
+                return true;
+            }
+
+            return false;
+        };
+
+        /**
+         * Mutating criticalSelectors Map
+         * @param ast
+         */
+        const gatherCriticalSelectors = (ast) => {
+            let _astRoot = {};
+            let media    = null;
+
+            // Root knot or media query
+            if (ast.stylesheet) {
+                _astRoot = ast.stylesheet;
+            } else if (ast.rules && ast.type === "media") {
+                _astRoot = ast;
+                media    = ast.media;
+            } else {
+                console.warn("Missing ast rules!!!", ast.type, ast.stylesheet);
+            }
+
+            const rules = _astRoot.rules;
+
+            for (let rule of rules) {
+                // Part of useful rule types
+                if (usedSelectorTypes.includes(rule.type)) {
+                    // If rule type is media -> rerun as rule
+                    if (rule.type === "media") {
+                        // Recursive process of media rule
+                        gatherCriticalSelectors(rule);
+                    } else { // IS ALWAYS RULE
+
+                        const selectors = rule.selectors || [];
+                        media           = media || "";
+
+                        // Key for identify
+                        const ruleKey = media + selectors.join();
+
+                        for (let selector of selectors) {
+
+                            // Check if selector is pure pseudo or a critical match
+                            // NOTE: Check if we are in trouble with doubled selectors with different content
+                            if (isPurePseudo(selector) || isSelectorCritical(selector)) {
+                                if (criticalSelectors.has(ruleKey)) {
+                                    const critSel = criticalSelectors.get(ruleKey);
+                                    if (!critSel.selectors.includes(selector)) {
+                                        critSel.selectors.push(selector);
+                                    }
+                                } else {
+                                    criticalSelectors.set(ruleKey, {
+                                        selectors: [selector],
+                                        media:     media, // Needed?
+                                        rule:      rule // Needed? maybe for doubled rules
+                                    });
+                                }
+                            }
+                        }
+                    }
+                } else {
+                    console.debug("DEBUG: UNPROCESSED RULE TYPE: " + rule.type);
+                }
+            }
+        };
+
+        gatherCriticalSelectors(sourceAst);
+
+        return resolve([...criticalSelectors]);
+
+    }).catch(error => {
+        console.error(error);
+    })
+};
