+++ lib/evaluation/extract_critical_without_css.js
@@ -1,1 +1,116 @@
-module.exports=e=>{const t=!!e.removePseudoSelectors,r=e.keepSelectors||[],n=(e.renderTimeout,!!e.dropKeyframes),c=window.innerHeight,o=t?/([^\s,\:\(])\:\:?(?!not)[a-zA-Z\-]{1,}(?:\(.*?\))?/g:/(.*)/,l=[],u=document.createTreeWalker(document,NodeFilter.SHOW_ELEMENT,{acceptNode:function(e){return NodeFilter.FILTER_ACCEPT}},!0);for(;u.nextNode();){const e=u.currentNode;if(r.find(t=>e.matches(t)))l.push(e);else{e.getBoundingClientRect().top<c&&l.push(e)}}const s=document.styleSheets,a=(e,t,r)=>e.filter((function(e){return e.matches(t.selectorText.replace(r,"$1"))})).length>0;return Array.prototype.map.call(s,(function(e){const t=e.rules||e.cssRules;return t?{sheet:e,rules:Array.prototype.map.call(t,(function(e){try{if(e instanceof CSSMediaRule){let t=e.rules||e.cssRules,r=Array.prototype.filter.call(t,(function(e){return a(l,e,o)})).map((function(e){return e.cssText})).reduce((function(e,t){return t+"\n"+e}),"");return r?"@media "+e.media.mediaText+" { "+r+"}":null}return e instanceof CSSStyleRule?(-1!==e.selectorText.indexOf(".c24-travel-main-cnt .c24-travel-extended-fieldset, .c24-travel-main-cnt .c24-travel-searchform-wrapper")&&console.log(e.cssText),a(l,e,o)?e.cssText:null):n&&(e instanceof CSSKeyframeRule||e instanceof CSSKeyframesRule)?"":e.cssText}catch(t){throw console.error("Bad CSS rule",e.selectorText),t}})).filter((function(e){return e}))}:null})).filter((function(e){return e&&e.rules.length>0})).map((function(e){return e.rules.join("")})).reduce((function(e,t){return t+e}),"").replace(/\n/g,"").replace(/content\: \"(.)\"/g,(function(e,t){return'content: "\\'+encodeURI(t).substr(2)+'"'}))};
+/**
+ * Used to get the current live critical css without any impact of css
+ *
+ * @param options
+ * @returns {*}
+ */
+module.exports = (options) => {
+    // ADJUSTMENTS
+    const removePseudoSelectors = !!options.removePseudoSelectors;
+    const keepSelectors         = options.keepSelectors || [];
+    const renderTimeout         = options.renderTimeout || 300;
+    const dropKeyframes         = !!options.dropKeyframes;
+
+    // innerHeight of window to determine if in viewport
+    const height        = window.innerHeight;
+    //Setup our Pseudo selector killer, view height and critical nodes
+    const removePseudo  = removePseudoSelectors ? /([^\s,\:\(])\:\:?(?!not)[a-zA-Z\-]{1,}(?:\(.*?\))?/g : /(.*)/;
+    const criticalNodes = [];
+
+    // Get a list of all the elements in the view.
+    const walker = document.createTreeWalker(
+        document,
+        NodeFilter.SHOW_ELEMENT,
+        {
+            acceptNode: function (node) {
+                return NodeFilter.FILTER_ACCEPT;
+            }
+        },
+        true
+    );
+
+    while (walker.nextNode()) {
+        const node = walker.currentNode;
+
+        const keep = keepSelectors.find(selector => {
+            // TODO: wildcard
+            // node.matches('[class*="te"]')
+            // node.matches('[id*="te"]')
+            // node.matches('[*="te"]')
+            return node.matches(selector);
+        });
+
+        if (keep) {
+            criticalNodes.push(node);
+        } else {
+            const rect = node.getBoundingClientRect();
+            if (rect.top < height) {
+                criticalNodes.push(node);
+            }
+        }
+    }
+
+    // Grab loaded stylesheets
+    const sheets = document.styleSheets;
+
+    const filterNodes = (nodes, rule, replace) => {
+        return nodes.filter(function (e) {
+            return e.matches(rule.selectorText.replace(replace, "$1"))
+        }).length > 0;
+    };
+
+    const outCss = Array.prototype.map.call(sheets, function (sheet) {
+        const rules = sheet.rules || sheet.cssRules;
+        // If there are rules
+        if (rules) {
+            return {
+                sheet: sheet,
+                rules: Array.prototype.map.call(rules, function (rule) { // Convert each CSSRule into a string
+                    try {
+                        if (rule instanceof CSSMediaRule) {
+                            let subRules = rule.rules || rule.cssRules;
+                            let css      = Array.prototype.filter.call(subRules, function (rule) {
+                                return filterNodes(criticalNodes, rule, removePseudo);
+                            }).map(function (rule) {
+                                return rule.cssText
+                            }).reduce(function (ruleCss, init) {
+                                return init + "\n" + ruleCss;
+                            }, "");
+                            return css ? ("@media " + rule.media.mediaText + " { " + css + "}") : null;
+
+                        } else if (rule instanceof CSSStyleRule) {
+
+                            if (rule.selectorText.indexOf(".c24-travel-main-cnt .c24-travel-extended-fieldset, .c24-travel-main-cnt .c24-travel-searchform-wrapper") !== -1) {
+                                console.log(rule.cssText);
+                            }
+
+                            return filterNodes(criticalNodes, rule, removePseudo) ? rule.cssText : null;
+
+                        } else if (dropKeyframes && (rule instanceof CSSKeyframeRule || rule instanceof CSSKeyframesRule)) {
+                            return "";
+                        } else {
+                            return rule.cssText;
+                        }
+                    } catch (e) {
+                        console.error("Bad CSS rule", rule.selectorText);
+                        throw e;
+                    }
+                }).filter(function (e) {
+                    return e;
+                })
+            }
+        } else {
+            return null;
+        }
+    }).filter(function (cssEntry) {
+        return cssEntry && cssEntry.rules.length > 0
+    }).map(function (cssEntry) {
+        return cssEntry.rules.join("");
+    }).reduce(function (css, out) {
+        return out + css
+    }, "");
+
+    return outCss.replace(/\n/g, "").replace(/content\: \"(.)\"/g, function (a, e) {
+        return "content: \"\\" + encodeURI(e).substr(2) + "\"";
+    });
+};
